refactor: v26.41.8 拆分巨石文件（第二轮）— 8 个文件拆为子模块
沿用 v26.41.7 的拆分模式，继续拆分 8 个超大文件：

- court_guarantee_api.py (1376行 → schemas + helpers + thin API)
- court_filing_api.py (1214行 → schemas + helpers + thin API)
- sales_dispute_api.py (1059行 → 4 个域路由 + factories + orchestrator)
- display_mixin.py (1037行 → display_format_mixin + trimmed mixin)
- automation_protocols.py (713行 → 4 个领域协议 + re-export facade)
- dashboard_service.py (718行 → dashboard_models + trimmed service)
- complete_defaults.py (797行 → JSON + 载入函数)
- contract_admin.py (692行 → contract_inlines + trimmed admin)

所有拆分保持外部导入路径不变，302 个单元测试全部通过。
