+++ README.md
@@ -1,4 +1,4 @@
-# 法穿 AI Copilot · V26.41.7
+# 法穿 AI Copilot · V26.41.8
 
 *由一线执业律师主导研发 · 开源 · 私有化部署 · 数据自持*
