+++ backend/apps/contracts/admin/mixins/display_mixin.py
@@ -1,27 +1,27 @@
 """
 Contract Admin - Display Mixin
 
-显示相关的方法:详情页视图、字段显示等.
+详情页视图、URL 路由等视图方法.
 """
 
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
+from typing import TYPE_CHECKING, Any
 
 from django.contrib import admin
 from django.core.exceptions import PermissionDenied
 from django.http import Http404, HttpRequest, HttpResponse
 from django.shortcuts import render
 from django.urls import URLPattern, path, reverse
-from django.utils.html import format_html
 from django.utils.translation import gettext_lazy as _
 
 from apps.core.exceptions import BusinessException, NotFoundError
 from apps.contracts.models.finalized_material import FinalizedMaterial
 
+from .display_format_mixin import ContractDisplayFormatMixin, _get_contract_display_service
+
 if TYPE_CHECKING:
-    from django.contrib.admin import ModelAdmin
     from django.db.models import Model
 
 logger = logging.getLogger("apps.contracts")
@@ -34,13 +34,6 @@ def _get_contract_admin_service() -> Any:
     return get_contract_admin_service()
 
 
-def _get_contract_display_service() -> Any:
-    """工厂函数获取合同显示服务"""
-    from apps.contracts.admin.wiring_admin import get_contract_display_service
-
-    return get_contract_display_service()
-
-
 def _get_contract_detail_reminders(contract: Any) -> list[dict[str, Any]]:
     from apps.core.interfaces import ServiceLocator
 
@@ -48,8 +41,8 @@ def _get_contract_detail_reminders(contract: Any) -> list[dict[str, Any]]:
     return reminder_service.export_contract_reminders_internal(contract_id=contract.id)
 
 
-class ContractDisplayMixin:
-    """合同 Admin 显示相关方法的 Mixin"""
+class ContractDisplayMixin(ContractDisplayFormatMixin):
+    """合同 Admin 视图方法的 Mixin（继承显示方法）"""
 
     if TYPE_CHECKING:
         admin_site: Any
@@ -57,124 +50,6 @@ class ContractDisplayMixin:
 
         def has_view_permission(self, request: HttpRequest, obj: Any = None) -> bool: ...
         def has_change_permission(self, request: HttpRequest, obj: Any = None) -> bool: ...
-
-    @admin.display(description=_("合同名称"), ordering="name")
-    def name_link(self, obj: Any) -> Any:
-        """生成指向详情页的合同名称链接"""
-        url = reverse("admin:contracts_contract_detail", args=[obj.pk])
-        return format_html('<a href="{}">{}</a>', url, obj.name)
-
-    @admin.display(description=_("主办律师"))
-    def get_primary_lawyer(self, obj: Any) -> Any:
-        """显示主办律师（使用 prefetch_related 数据避免 N+1）"""
-        for assignment in obj.assignments.all():
-            if assignment.is_primary:
-                lawyer = assignment.lawyer
-                return lawyer.real_name or lawyer.username
-        return "-"
-
-    def _get_primary_lawyer_obj(self, obj: Any) -> Any:
-        """返回主办律师对象（供详情页模板使用）"""
-        for assignment in obj.assignments.all():
-            if assignment.is_primary:
-                return assignment.lawyer
-        return None
-
-    @admin.display(description=_("主办律师"))
-    def get_primary_lawyer_display(self, obj: Any) -> Any:
-        """详情页显示主办律师(只读)"""
-        from apps.contracts.admin.wiring_admin import get_contract_assignment_query_service
-
-        service = get_contract_assignment_query_service()
-        assignment = service.get_primary_lawyer(obj.pk)
-        if assignment:
-            lawyer = assignment.lawyer
-            name = lawyer.real_name or lawyer.username
-            return f"{name} (ID: {lawyer.id})"
-        return _("无")
-
-    @admin.display(description=_("律所OA链接"))
-    def law_firm_oa_link_display(self, obj: Any) -> Any:
-        """显示合同所属律所的 OA 登录链接（可点击）。"""
-        from apps.oa_filing.services.script_executor_service import SUPPORTED_SITES
-        from apps.organization.models import AccountCredential
-
-        law_firm_ids: list[int] = []
-        seen: set[int] = set()
-        for assignment in obj.assignments.select_related("lawyer").all():
-            lawyer = getattr(assignment, "lawyer", None)
-            law_firm_id = getattr(lawyer, "law_firm_id", None)
-            if not law_firm_id or law_firm_id in seen:
-                continue
-            seen.add(int(law_firm_id))
-            law_firm_ids.append(int(law_firm_id))
-
-        if not law_firm_ids:
-            return _("未配置")
-
-        credential = (
-            AccountCredential.objects.filter(
-                lawyer__law_firm_id__in=law_firm_ids,
-                site_name__in=SUPPORTED_SITES,
-            )
-            .exclude(url__isnull=True)
-            .exclude(url="")
-            .order_by("id")
-            .first()
-        )
-
-        if not credential:
-            return _("未配置")
-
-        return format_html(
-            '<a href="{}" target="_blank" rel="noopener noreferrer">{}</a>',
-            credential.url,
-            _("打开OA系统"),
-        )
-
-    @admin.display(description=_("建档编号"))
-    def filing_number_display(self, obj: Any) -> Any:
-        """显示建档编号(只读)
-
-        如果合同已有建档编号,显示编号;否则显示"未生成".
-
-        Requirements: 1.1, 1.2, 3.1
-        """
-        if obj and obj.filing_number:
-            return obj.filing_number
-        return _("未生成")
-
-    @admin.display(description=_("匹配的合同模板"))
-    def get_matched_template_display(self, obj: Any) -> Any:
-        """显示匹配的合同模板
-
-        Requirements: 1.4
-        """
-        if not obj or not obj.pk:
-            return _("请先保存合同")
-
-        try:
-            display_service = _get_contract_display_service()
-            return display_service.get_matched_document_template(obj)
-        except (BusinessException, RuntimeError, Exception) as e:
-            logger.error("获取合同 %s 匹配模板失败: %s", obj.id, e, exc_info=True)
-            return _("查询失败")
-
-    @admin.display(description=_("匹配的文件夹模板"))
-    def get_matched_folder_templates_display(self, obj: Any) -> Any:
-        """显示匹配的文件夹模板
-
-        Requirements: 7.1
-        """
-        if not obj or not obj.pk:
-            return _("请先保存合同")
-
-        try:
-            display_service = _get_contract_display_service()
-            return display_service.get_matched_folder_templates(obj)
-        except (BusinessException, RuntimeError, Exception) as e:
-            logger.error("获取合同 %s 匹配文件夹模板失败: %s", obj.id, e, exc_info=True)
-            return _("查询失败")
 
     def get_urls(self) -> Any:
         """添加自定义 URL 路由"""
