+++ backend/apps/sales_dispute/api/sales_dispute_api_factories.py
@@ -1,0 +1,102 @@
+"""买卖纠纷计算 API — 共享工厂函数与工具。"""
+
+from __future__ import annotations
+
+from datetime import date
+from typing import Any
+
+
+def _get_interest_calculator() -> Any:
+    from apps.sales_dispute.services.interest_calculator_service import InterestCalculatorService
+
+    return InterestCalculatorService()
+
+
+def _get_cost_benefit_service() -> Any:
+    from apps.sales_dispute.services.cost_benefit_service import CostBenefitService
+
+    return CostBenefitService()
+
+
+def _get_lpr_rate_service() -> Any:
+    from apps.sales_dispute.services.lpr_rate_service import LprRateService
+
+    return LprRateService()
+
+
+def _get_case_assessment_service() -> Any:
+    from apps.sales_dispute.services.case_assessment_service import CaseAssessmentService
+
+    return CaseAssessmentService()
+
+
+def _get_limitation_calculator() -> Any:
+    from apps.sales_dispute.services.limitation_calculator_service import LimitationCalculatorService
+
+    return LimitationCalculatorService()
+
+
+def _get_jurisdiction_analyzer() -> Any:
+    from apps.sales_dispute.services.jurisdiction_analyzer_service import JurisdictionAnalyzerService
+
+    return JurisdictionAnalyzerService()
+
+
+def _get_strategy_recommender() -> Any:
+    from apps.sales_dispute.services.litigation_strategy_service import LitigationStrategyService
+
+    return LitigationStrategyService()
+
+
+def _get_collection_workflow() -> Any:
+    from apps.sales_dispute.services.collection_workflow_service import CollectionWorkflowService
+
+    return CollectionWorkflowService()
+
+
+def _get_collection_reminder() -> Any:
+    from apps.sales_dispute.services.collection_reminder_service import CollectionReminderService
+
+    return CollectionReminderService()
+
+
+def _get_lawyer_letter_generator() -> Any:
+    from apps.sales_dispute.services.lawyer_letter_generator_service import LawyerLetterGeneratorService
+
+    return LawyerLetterGeneratorService()
+
+
+def _get_reconciliation_generator() -> Any:
+    from apps.sales_dispute.services.reconciliation_generator_service import ReconciliationGeneratorService
+
+    return ReconciliationGeneratorService()
+
+
+def _get_settlement_generator() -> Any:
+    from apps.sales_dispute.services.settlement_generator_service import SettlementGeneratorService
+
+    return SettlementGeneratorService()
+
+
+def _get_execution_doc_generator() -> Any:
+    from apps.sales_dispute.services.execution_doc_generator_service import ExecutionDocGeneratorService
+
+    return ExecutionDocGeneratorService()
+
+
+def _get_dashboard_service() -> Any:
+    from apps.sales_dispute.services.dashboard_service import DashboardService
+
+    return DashboardService()
+
+
+def _resolve_date_range(
+    start_date: date | None,
+    end_date: date | None,
+) -> tuple[date, date]:
+    """未提供日期时默认当前自然年"""
+    today = date.today()
+    return (
+        start_date or date(today.year, 1, 1),
+        end_date or date(today.year, 12, 31),
+    )
